# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Omnichannel gateway, campaign engine, Direct Universal SIP Trunking PBX, MCP (Model Context Protocol), Unified RAG Router, Recall.ai Meeting Bots \u0026 Evolution Go Proxy",
        "title": "Patter Engine Gateway \u0026 Campaign Engine API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/lynxflow/patter-go"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/gateway/voice/call": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gateway - Voice"
                ],
                "summary": "Initier un appel vocal entrant / sortant avec Agent IA",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID du Tenant White-Label",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Paramètres de l'appel",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.VoiceCallRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/core.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.VoiceCallResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "core.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "service.VoiceCallRequest": {
            "type": "object",
            "required": [
                "from_number",
                "to_number"
            ],
            "properties": {
                "agent_id": {
                    "type": "string"
                },
                "from_number": {
                    "type": "string"
                },
                "prompt": {
                    "type": "string"
                },
                "to_number": {
                    "type": "string"
                }
            }
        },
        "service.VoiceCallResponse": {
            "type": "object",
            "properties": {
                "call_id": {
                    "type": "string"
                },
                "from_number": {
                    "type": "string"
                },
                "status": {
                    "description": "\"QUEUED\", \"RINGING\", \"IN_PROGRESS\"",
                    "type": "string"
                },
                "to_number": {
                    "type": "string"
                }
            }
        },
...
    }
}